requirements.txt
ortools
sqlalchemy
fastapi
uvicorn
jinja2
psycopg2-binary
python-dotenv
pydantic
groq
itsdangerous

